move Settings to its own module
..and add some tests for it
